+++ fabric/src/main/java/smartin/miapi/fabric/MiapiFabric.java
@@ -3,19 +3,15 @@ package smartin.miapi.fabric;
 import dev.architectury.platform.Platform;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.item.v1.EnchantmentEvents;
 import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
-import net.fabricmc.fabric.api.tag.FabricTagKey;
 import net.fabricmc.fabric.api.util.TriState;
-import net.fabricmc.fabric.impl.tag.convention.v2.TagRegistration;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.world.entity.ai.attributes.RangedAttribute;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
 import smartin.miapi.Environment;
 import smartin.miapi.Miapi;
@@ -43,11 +39,6 @@ public class MiapiFabric implements ModInitializer {
             MiapiEvents.DEFAULT_LOOT_FUNCTIONS.invoker().adjust(functions);
             functions.forEach(tableBuilder::apply);
         });
-        Block block;
-        FabricTagKey fabricTagKey;
-        FabricTagProvider.ItemTagProvider tagProvider;
-        TagRegistration tagRegistration;
-
         //DATA
         if (Environment.isClient()) {
             MiapiClientFabric.setupClient();