+++ CHANGELOG.md
@@ -13,5 +13,6 @@
 - fixed bug related to rendering multiple banners at once (having 2 armpatches now works)
 - fixed issues with crossbow animations desyncing/playing wrong
 - fixed bug where a load fail module extension would also unload the module
- ### Editor
+- fixed global loot modularisation  
+### Editor
 - reloading now closes ui to prevent crashes